Hide market insights when empty, and better job of hiding shows when empty
