+++ src/Apps/__stories__/Apps.story.tsx
@@ -17,12 +17,8 @@ storiesOf("Apps", module)
     return (
       <StorybooksRouter
         routes={artistRoutes}
-        initialRoute="/artist2/andy-warhol"
-        initialState={{
-          mediator: {
-            trigger: x => x,
-          },
-        }}
+        initialRoute="/artist2/shara-hughes"
+        initialState={{ mediator: { trigger: x => x } }}
       />
     )
   })
